pyproject.toml
[project]
name = "donna"
version = "0.1.0"
description = "AI Personal Assistant — active task management, scheduling, and autonomous sub-agents"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    # Core async
    "aiosqlite>=0.20.0",
    "aiohttp>=3.9.0",
    # External data sources
    "feedparser>=6.0.10,<7",
    # LLM providers
    "anthropic>=0.40.0",
    # Discord bot
    "discord.py>=2.3.0",
    # Google APIs
    "google-api-python-client>=2.100.0",
    "google-auth-oauthlib>=1.2.0",
    "google-auth-httplib2>=0.2.0",
    # Database & migrations
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    # Structured logging
    "structlog>=24.1.0",
    # Task deduplication
    "rapidfuzz>=3.5.0",
    # Schema validation
    "jsonschema>=4.20.0",
    # Configuration
    "pyyaml>=6.0.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # REST API (Phase 4 Flutter backend)
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "argon2-cffi>=23.1.0",
    # MCP server (LLM-facing tools)
    "fastmcp>=3.0.0",
    # Supabase sync
    "supabase>=2.0.0",
    # Automation subsystem (Phase 5)
    "croniter>=2.0.0",
    # Twilio (SMS/Voice)
    "twilio>=9.0.0",
    # Interactive CLI
    "questionary>=2.0.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "uuid6>=2024.1.12",
    # Skill system (Phase 1)
    "sentence-transformers>=3.0.0",
    "numpy>=1.26.0",
    # Skill system (Phase 2)
    "httpx>=0.27.0",
    # Article extraction
    "trafilatura>=1.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "aioresponses>=0.7.6",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.26.0",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-literate-nav>=0.6.0",
    "mkdocs-section-index>=0.3.8",
    "mkdocs-awesome-pages-plugin>=2.9.0",
]

[project.scripts]
donna = "donna.cli:main"

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (may use real SQLite)",
    "llm: Tests that call LLM APIs (costs money)",
    "slow: Slow tests (skip in CI by default)",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
allowed-confusables = ["—", "–", "’", "×"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = [
    "fastapi.Body",
    "fastapi.Depends",
    "fastapi.File",
    "fastapi.Form",
    "fastapi.Header",
    "fastapi.Path",
    "fastapi.Query",
]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "aioresponses>=0.7.6",
    "ruff>=0.4.0",
    "mypy>=1.8.0",
]
